+++ src/main/java/org/mithwick93/accountable/dal/repository/AssetRepository.java
@@ -8,8 +8,8 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AssetRepository extends JpaRepository<Asset, Long> {
+public interface AssetRepository extends JpaRepository<Asset, Integer> {
     List<Asset> findAllByUserId(int userId);
 
-    Optional<Asset> findByIdAndUserId(Long assetId, int userId);
+    Optional<Asset> findByIdAndUserId(int assetId, int userId);
 }
